More type fixes for windows
More fixes to int types to fix windows build issues.

+++ arrows/tests/test_image.h
@@ -86,12 +86,12 @@ populate_vital_image(kwiver::vital::image& img)
 }
 
 // Parameters for common test of get_image function
-constexpr size_t full_width = 60;
-constexpr size_t full_height = 40;
-constexpr size_t cropped_width = 30;
-constexpr size_t cropped_height = 20;
-constexpr size_t x_offset = 10;
-constexpr size_t y_offset = 5;
+constexpr unsigned int full_width = 60;
+constexpr unsigned int full_height = 40;
+constexpr unsigned int cropped_width = 30;
+constexpr unsigned int cropped_height = 20;
+constexpr unsigned int x_offset = 10;
+constexpr unsigned int y_offset = 5;
 
 // ----------------------------------------------------------------------------
 // helper function to generate common test of get_image funcion
